Full Stack Open Part 2d: Exercises 2.12 - 2.13: Refactor to move components into src. Add db.json file. Add json-server and axios dependencies to package.json. Create services directory and personServices module. Create get and add functions in personServices, which return a promise passing the response data. Import personsService into App component. Refactor useEffect to call personsService.get to initialize the app with the persons data from the server. Refactor handleAddPerson to call personsService.add. Chain setPersons to personsService.add and .get using .then to update state when promise resolves. Use .catch for error handling
