+++ exercises/parts-2-a--2-e/phonebook/src/AddPerson.jsx
@@ -1,4 +1,4 @@
-import Input from './Input'
+import Input from './Input.jsx'
 
 const AddPerson = ({newName, newNumber, handleNameInput, handleNumberInput, handleAddPerson}) => {
     return (
