+++ exercises/parts-2-a--2-e/phonebook/src/App.jsx
@@ -1,22 +1,28 @@
-import { useState } from 'react'
+import { useState, useEffect } from 'react'
 
-import AddPerson from '../AddPerson'
-import Search from '../Search'
-import TelephoneDirectory from '../TelephoneDirectory'
+import AddPerson from './AddPerson.jsx'
+import Search from './Search.jsx'
+import TelephoneDirectory from './TelephoneDirectory.jsx'
+
+import personsService from './services/personsService.js'
 
 const App = () => {
 
-    const [persons, setPersons] = useState([
-        { name: 'Arto Hellas', number: '040-123456', id: 1 },
-        { name: 'Ada Lovelace', number: '020-987654', id: 2 },
-        { name: 'Annie Easley', number: '401-223-1234', id: 3 },
-        { name: 'Dorothy Vaughan', number: '774-123-0987', id: 4 },
-        { name: 'Grace Hopper', number: '508-332-2233', id: 5 }
-    ])
-
-    const [searchString, setSearchString] = useState('')
     const [newName, setNewName] = useState('')
     const [newNumber, setNewNumber] = useState('')
+    const [persons, setPersons] = useState([])
+    const [searchString, setSearchString] = useState('')
+
+    useEffect(() => {
+        personsService
+            .get()
+            .then(personsData => {
+                setPersons(personsData)
+            })
+            .catch((error) => {
+                console.log('promise rejected', error)
+            })
+    }, [])
 
     const searchResults = (searchString) => {
         return persons.filter(
@@ -33,11 +39,17 @@ const App = () => {
     const handleAddPerson = (event) => {
         event.preventDefault()
         if (!matchingPersonsCount(newName)) {
-            setPersons(persons.concat({
-                name: newName,
-                number: newNumber,
-                id: persons.length + 1
-            }))
+            personsService
+                .add({
+                    name: newName,
+                    number: newNumber,
+                })
+                .then(newPersonData => {
+                    setPersons(persons.concat(newPersonData))
+                })
+                .catch((error) => {
+                    console.log('promise rejected', error)
+                })
         } else {
             alert(`${newName.toLowerCase()} is already added to phonebook`)
         }
