+++ exercises/parts-2-a--2-e/phonebook/src/Search.jsx
@@ -1,4 +1,4 @@
-import Input from './Input'
+import Input from './Input.jsx'
 
 const Search = ({searchString, handleSearch}) => {
     return (
