+++ exercises/parts-2-a--2-e/phonebook/src/db.json
@@ -1,0 +1,49 @@
+{
+  "persons": [
+    {
+      "name": "Arto Hellas",
+      "number": "040-123456",
+      "id": "1"
+    },
+    {
+      "name": "Ada Lovelace",
+      "number": "39-44-5323523",
+      "id": "2"
+    },
+    {
+      "name": "Dan Abramov",
+      "number": "12-43-234345",
+      "id": "3"
+    },
+    {
+      "name": "Mary Poppendieck",
+      "number": "39-23-6423122",
+      "id": "4"
+    },
+    {
+      "name": "Grace Hopper",
+      "number": "11-22-33445566",
+      "id": "5"
+    },
+    {
+      "id": "3cc2",
+      "name": "Marry Kenneth Keller",
+      "number": "444-333-5555"
+    },
+    {
+      "id": "8be3",
+      "name": "Gladys West",
+      "number": "222233---000022-21"
+    },
+    {
+      "id": "cebe",
+      "name": "Frances E. Allen",
+      "number": "662-887-0099"
+    },
+    {
+      "id": "2767",
+      "name": "Linus Torvalds",
+      "number": "888--22=11001"
+    }
+  ]
+}
